Remove "comfy margin"
The original concept was that it makes it easier for the user
to recognize when a view scrolled to beginning or end by leaving
an empty margin when that happens, but in reality this messes with
the scrolling and sync for very small views, like one row tall views,
the whole content can go out of view.
It's just easier to remove this comfy margin to simplify implementation
of keeping content in view.